totally damn brokern.
git-svn-id: file:///mnt/sdd1/kens/svnroot/trunk/coop@1180 f16cc87a-5815-0410-8a9a-9fe0dcc0200c

+++ web/COOP/Form.php
@@ -322,7 +322,7 @@ class coopForm extends CoopObject
 			print $this->page->confessArray($vars, 
 									  'CoopForm::processAddRemove(vars)', 1);
 
-			//$this->obj->debugLevel(2);
+			//$this->obj->debugLevel(0); 
 			// for everything in crosslinks, 
 			foreach($this->obj->fb_crossLinks as $la){
 				$tf = $la['toField'];
@@ -340,9 +340,14 @@ class coopForm extends CoopObject
 				
 				// yeah, array_diff is the long way with db thrashing.
 				// but i want clear, easily-debugged code
+				$this->obj->{$this->pk} = $this->id;
 				$indb = $this->checkCrossLinks($mt,$ft);
+				$this->page->confessArray($indb, 
+										  'CoopForm::processCrossLinks(indb)');
 				$toSave = array_diff($vars[$tf], $indb);
-				$toDelete = array_diff($indb, $vars[$tf]);
+				if(count($vars[$tf]) < count($indb)){
+					$toDelete = array_diff($indb, $vars[$tf]);
+				}
 				$this->page->confessArray($toSave, 
 										  'CoopForm::processsCrossLinks(save)', 
 										  2);
@@ -353,20 +358,24 @@ class coopForm extends CoopObject
 
 				//$this->obj->debugLevel(2);
 				// save
-				foreach($toSave as $saveme){
-					$mid =& new CoopObject(&$this->page, $mt, &$this);
-					$mid->obj->$tf = $saveme;
-					$mid->obj->$nk = $this->id;
-					$mid->obj->insert();
+				if(is_array($toSave)){
+					foreach($toSave as $saveme){
+						$mid =& new CoopObject(&$this->page, $mt, &$this);
+						$mid->obj->$tf = $saveme;
+						$mid->obj->$nk = $this->id;
+						$mid->obj->insert();
+					}
 				}
 				
 				// delete
-				foreach($toDelete as $killme){
-					$mid =& new CoopObject(&$this->page, $mt, &$this);
-					$mid->obj->$tf = $killme;
-					$mid->obj->$nk = $this->id;
-					$mid->obj->limit(1);
-					$mid->obj->delete();
+				if(is_array($toDelete)){
+					foreach($toDelete as $killme){
+						$mid =& new CoopObject(&$this->page, $mt, &$this);
+						$mid->obj->$tf = $killme;
+						$mid->obj->$nk = $this->id;
+						$mid->obj->limit(1);
+						$mid->obj->delete();
+					}
 				}
 
 			}
@@ -381,7 +390,6 @@ class coopForm extends CoopObject
 				$ft = $la['toTable'];
 				$nk = $this->backlinks[$mt];
 
-				$incl = $this->checkCrossLinks($mt, $ft);
 				
 				//duplication of selectoptions
 				$this->page->debug > 3 && $this->obj->debugLevel(2);
@@ -404,17 +412,20 @@ class coopForm extends CoopObject
 										$far->title(),
 										$options);
 
-			//confessArray($incl,'included');
 
-			$this->form->setDefaults(
-				array($tf =>
-					  isset(
-						  $_REQUEST['_qf__' . 
-									$this->form->_attributes['name']]) ?
-					  $_REQUEST[$tf] :
-					  $incl));
-
+				$this->obj->{$this->pk} = $this->id; // for checkcrosslinks
+				$incl = $this->checkCrossLinks($mt, $ft);
+				confessArray($incl, 'incl');
+				$this->form->setDefaults(
+					array($tf =>
+						  $this->is_submitted ? $_REQUEST[$tf] : $incl));
+				
 			}
+		}
+
+	function is_submitted()
+		{
+			return isset($_REQUEST['_qf__' . $this->form->_attributes['name']]);
 		}
 
 } // END COOP FORM CLASS
